requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
opencv-python>=4.8.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
websockets>=12.0
ultralytics>=8.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.5.0
